scripts/vcpkg-prune: paginate list fetches

+++ scripts/vcpkg-prune.py
@@ -1,5 +1,6 @@
 #!/usr/bin/env python3
 
+import itertools
 import os
 
 import requests
@@ -13,22 +14,29 @@ HEADERS = {
     'X-GitHub-Api-Version': '2026-03-10',
 }
 
-resp = requests.get(
-    f'https://api.github.com/orgs/{GH_USERNAME}/packages?package_type=nuget',
-    headers=HEADERS,
-)
-resp.raise_for_status()
-for pkg in sorted(resp.json(), key=lambda pkg: pkg['name']):
-    if not pkg['name'].startswith(PACKAGE_PREFIX):
-        continue
-    resp = requests.get(f'{pkg["url"]}/versions', headers=HEADERS)
+for page in itertools.count(1):
+    resp = requests.get(
+        f'https://api.github.com/orgs/{GH_USERNAME}/packages?package_type=nuget&page={page}',
+        headers=HEADERS,
+    )
     resp.raise_for_status()
-    versions = sorted(resp.json(), key=lambda ver: ver['created_at'])
-    for ver in versions[:-3]:
-        print(
-            f'Deleting {pkg["name"]} {ver["name"]} created at {ver["created_at"]}'
-        )
-        # versions with more than 5k downloads cannot be deleted
-        resp = requests.delete(ver['url'], headers=HEADERS)
-        if resp.status_code != 204:
-            print('...failed:', resp.text)
+    packages = resp.json()
+    if not packages:
+        break
+    for pkg in packages:
+        if not pkg['name'].startswith(PACKAGE_PREFIX):
+            continue
+        while True:
+            resp = requests.get(f'{pkg["url"]}/versions', headers=HEADERS)
+            resp.raise_for_status()
+            prune = sorted(resp.json(), key=lambda ver: ver['created_at'])[:-3]
+            if not prune:
+                break
+            for ver in prune:
+                print(
+                    f'Deleting {pkg["name"]} {ver["name"]} created at {ver["created_at"]}'
+                )
+                # versions with more than 5k downloads cannot be deleted
+                resp = requests.delete(ver['url'], headers=HEADERS)
+                if resp.status_code != 204:
+                    print('...failed:', resp.text)
